+++ src/annbatch/__init__.py
@@ -2,20 +2,16 @@ from __future__ import annotations
 
 from importlib.metadata import version
 
-from . import abc, types
+from . import abc, samplers, types
 from .io import DatasetCollection, write_sharded
 from .loader import Loader
-from .samplers import ChunkSampler, DistributedChunkSampler, RandomSampler, SequentialSampler
 
 __version__ = version("annbatch")
 
 __all__ = [
     "Loader",
-    "ChunkSampler",
-    "DistributedChunkSampler",
-    "RandomSampler",
-    "SequentialSampler",
     "DatasetCollection",
+    "samplers",
     "types",
     "write_sharded",
     "abc",